JD02-03 - currentCashierCount implement as AtomicInteger

+++ src/by/it/marchenko/jd02_03/services/StoreWorker.java
@@ -215,9 +215,8 @@ public class StoreWorker extends Thread implements StoreAction {
     }
 
     @Override
-    public int getCurrentCashierCount() {
-        //return currentCashierCount;
-        return currentCashierCount.intValue();
+    public AtomicInteger getCurrentCashierCount() {
+        return currentCashierCount;
     }
 
     @Override
